Lots of clean up and issue fixing for the sample problems.
Fix warnings and errors in many of the sample problems, and clean up the
categories. Note that algebra and trigonometry should not be
categories. Those are subject areas.

The documentation in the problems has been updated for consistency,
readability, and correctness.

Cases where new PGML constructs could be used have been implemented
(tables, images, tags, and such).