Add a sine wave tool to the graphtool.
The tool can be added by setting `availableTools => [ 'SineWaveTool' ]`
in the `->with` call when creating a `GraphTool` object. Of course other
tools can be added in the list as well.  The tool works by plotting
three points.  The first point determines the phase shift (or x
translation) and y translation.  The second point determines the
amplitude.  Note that point can only be moved horizontally, and is
forced to always stay on the same horizontal line as the first point.
The third point determines the amplitude. That point can only be moved
vertically, and is forced to stay at the peak of the first crest of the
sine wave.  So if `(x1, y1)` is the first point, `x2` the x-coordinate
of the second point, and `y3` the y-coordinate of the third point, then
the function is `f(x) = y3 * sin(2 * pi / x2 * (x - x1)) + y1`.

There are several new GraphTool options added to make it so that the
axis labels can be multiples of `pi`.  They are `scaleX`, `scaleY`,
`scaleSymbolX`, and `scaleSymbolY`.  The "scale" options are the scale
of the ticks on the x and y axes. That is the distance between two
successive ticks on the axis (including both major and minor ticks).
The "scaleSymbol" options are the scale symbols for the ticks on the x
and y axes. The labels on the axis will be shown as multiples of this
symbol.

These options can be used in combination to show tick labels at
multiples of pi by using the settings `scaleX => pi->value` and
`scaleSymbolX => "~~x{03C0}"`.  Note that value of `scaleSymbolX` is the
unicode pi symbol, and that `~~` is the PG escape character (normally a
backslash in Perl).

The coordinate hints in the lower right corner have also been updated to
use these options if set.

There is a general and minor change to some other graph objects.  The
"down", "up", and "drag" handlers are no longer attached to static graph
object points. For most objects this doesn't matter because the points
are hidden. So the points can't be clicked on or dragged in any case.
However, for the "fill" graph object the paint bucket icon is still
visible for static fill objects, and then clicking on the icon you see
the mouse cursor disappear.  The icon still can't be dragged because it
is fixed, but the cursor disappearing sort of indicates that the icon
has been grabbed for a drag which is somewhat counter intuitive when you
can't actually do so. These handlers aren't needed for static objects in
any case.