Also make the axis tick labels be shown according to the `coordinateHintsType` settings.
This reworks the coordinate hints code considerably.  The JSXGraph
`formatLabelText` method of the default axes ticks is overriden so that
tick labels can be formatted according to our `coordinateHintsType`
settings.  Note that there is a JSXGraph tick label `toFraction` option
that is similar and is used in the default `formatLabelText` method, but
it always uses mixed numbers. Furthermore, the `scaleSymbol` is included
in what is typeset by MathJax (the default method does not do this).
This looks better and also means that instead of needing to use the
unicode symbol for pi, the LaTeX `\pi` command can be used instead. Note
that MathJax actually supports the unicode symbol and typesets it the
same as `\pi`.

So if the `coordinateHintsType` setting is 'mixed' then both the
coordinate hints shown in the lower right corner of the graph (or upper
left for number lines) and the labels on the axes will be shown as mixed
numbers, and if the setting is 'fraction' then both will be displayed as
improper fractions, and with the default 'decimal' setting both will be
decimals.

Note that the `useMathJax: true` and `display: 'html'` settings are now
set for the axes tick labels so that fractions are displayed typeset by
MathJax.

Note that this is stated above for the `coordinateHintsType` setting,
but really applies to the `coordinateHintsTypeX` and
`coordinateHintsTypeY` settings as well for per coordinate and axis
behavior.

Also note that now the `coordinateHintsType` settings are perhaps not
appropriately named since they apply to more than just the coordinate
hints, and also apply to axis labels. But I don't feel like adding
another setting for this and consistency between the hint coordinates
and axis labels is better.

The JSXGraph `generateLabelText` method of the axes ticks is also
overridden so that zero is also typeset using MathJax.  The default
JSXGraph method never does this even if the other labels are typeset
using MathJax and gives an inconsistent appearance in that case.