Add a way for a graph object to add to the help.
A graph object can only add to the help when it is selected.

The help is also updated when focus moves within the object to different
defining points.  So the graph object can give a different message
depending on which point is focused.  The sine wave tool uses this to
let the user know restrictions on how the focused point can move.