Remove "positive violations" and invert score.
Now all violations are really violations.  There is no need for the
"positive violations".  The score is now purely a badness score.  The
higher the score, the more work is needed to fix the problem.  A score
of 0 means the problem has no issues.

The "positive violations" were really going to be a problem if the
intent of the score is to obtain a measure of how much work is needed to
fix a problem and make it conform to current best-practices in problem
authoring.  For example, a problem could do quite a bit wrong but have a
custom checker.  The custom checker score was really high, and so that
would offset the things done wrong and it might end up with a score that
is the same as for a problem that only does a few things wrong, but
doesn't have a custom checker.  So that first problem with the custom
checker really needs a lot of work, but the second problem without the
checker only needs trivial fixes.  By only having a badness score you
get a much clearer measure of how much is needed to update a problem to
conform with current best-practices.