Add a workflow to generate and publish PG POD and sample problem documentation.
This is intended to replace the openwebwork/pg-docs repository.  The
approach in that repository has some issues.

* The workflows must be manually triggered in sequence.
* It isn't tied to the PG repository and so can become and in fact is
  now out of date.
* It unnecessarily commits the generated documentation files to a
  repository as well as publishing them on GitHub pages.
* It uses the `bin/dev_script/generate-ww-pg-pod.pl` script from the
  webwork2 repository to generate the PG and webwork2 POD, and the
  webwork2 POD should not be on a site that is intended for PG
  documentation.

The single workflow added in this pull request runs anytime anything is
merged into the main branch.  It can also be manually triggered but that
generally won't be needed.  This means that anytime a new version of PG
is released the documentation will automatically be updated for that
release.

The workflow generates the PG POD and sample problem documentation and
publishes the documentation to the PG GitHub pages site at
https://openwebwork.github.io/pg.  The generated documentation is not
committed to a repository.  It is published to the pages site, and that
is enough.

A new `bin/generate-pg-pod.pl` script is added that generates only the
PG POD.  The generated html is tailored for the PG documentation on
GitHub pages.  There is some minor redundancy as there are some files
that are now in both the webwork2 and PG repositories.  However, it
certainly would not be a good idea to have the PG repository depend on
the webwork2 repository.