Clean up of the sample problem parser code.
Move the `SampleProblemParser` module into the `WeBWorK::PG` namespace.
All of the other modules in the root namespace in the PG repository are
directly for problem rendering except this one.  It is out of place.  So
it is now the `WeBWorK::PG::SampleProblemParser` module.

Change the options for the `parser-problem-doc.pl` script and the
`WeBWorK::PG::SampleProblemParser` module.  Using underscores for
options is annoying.  Underscores require an extra keyboard button press
(the shift key), and underscore are not standard for command line
options.  Also the `pod_root` and `pg_doc_home` options are badly named.
The `pod_root` option seems like it should be the directory location for
the POD, but it is a URL.  So make that clear by renaming it to
`pod-base-url`.  The `pg_doc_home` is not the URL for the PG
documentation home, but for the sample problem base URL.  So rename that
to `sample-problem-base-url` (and its shortcut from "h" to "s").  Also
change the corresponding variable names in the `SampleProblemParser.pm`
module.

The workflow is updated for the changes.

Webwork2 also needs to be updated for these changes.