Implement several options for the GraphTool in hardcopy that should have been implemented before.
If the `scaleX`, `scaleSymbolX`, `scaleY`, `scaleSymbolY`, or
`coordinateHintsType` options are used then the axis ticks and labels
are adjusted in the JSXGraph HTML output for the GraphTool, but not in
hardcopy.

This was an oversight that I didn't do this when these options were
implemented.

I also added minor ticks to the hardcopy.  Those were also never
implemented.

The documentation for the `scaleX` and `scaleY` options was also
corrected.  I misstated the way that this is handled by JSXGraph.